feat(table): +files and folders management, +styles +types files
