+++ packages/mantine/src/components/table/table-actions/TableActions.styles.ts
@@ -1,0 +1,8 @@
+import {createStyles} from '@mantine/core';
+
+export default createStyles((theme) => ({
+    root: {},
+    wrapper: {
+        display: 'inline-flex',
+    },
+}));
