+++ packages/mantine/src/components/table/table-actions/TableActions.tsx
@@ -1,40 +1,10 @@
-import {createStyles, DefaultProps, Grid, Group, Selectors} from '@mantine/core';
+import {Grid, Group} from '@mantine/core';
 import {ReactElement, ReactNode} from 'react';
 
 import {TableComponentsOrder} from '../Table.styles';
 import {useTable} from '../TableContext';
-
-const useStyles = createStyles((theme) => ({
-    root: {},
-    wrapper: {
-        display: 'inline-flex',
-    },
-}));
-
-type TableActionsStylesNames = Selectors<typeof useStyles>;
-
-interface TableActionsProps<T> extends DefaultProps<TableActionsStylesNames> {
-    /**
-     * Function that return components for the selected row or selected rows when multi row selection is enabled
-     *
-     * @param datum the data of the selected row(s)
-     * @example
-     * <Table.Actions<MyType>>
-     *     {(datum: MyType) => (
-     *         <Button
-     *             component={Link}
-     *             to={`edit/${datum.id}`}
-     *             leftIcon={<EditSize24Px />}
-     *             variant="subtle"
-     *             color="navyBlue.8"
-     *         >
-     *             Edit
-     *         </Button>
-     *     )}
-     * </Table.Actions>
-     */
-    children: ((datum: T) => ReactNode) | ((data: T[]) => ReactNode);
-}
+import useStyles from './TableActions.styles';
+import {TableActionsProps} from './TableActions.types';
 
 export const TableActions = <T,>({
     children,
