+++ packages/mantine/src/components/table/table-date-range-picker/TableDateRangePicker.styles.ts
@@ -1,0 +1,7 @@
+import {createStyles} from '@mantine/core';
+
+export default createStyles((theme) => ({
+    root: {},
+    wrapper: {},
+    label: {},
+}));
