+++ packages/mantine/src/components/table/table-filter/TableFilter.styles.ts
@@ -1,0 +1,11 @@
+import {createStyles} from '@mantine/core';
+
+export default createStyles((theme) => ({
+    root: {},
+    wrapper: {
+        marginBottom: '0 !important',
+    },
+    empty: {
+        color: theme.colors.gray[4],
+    },
+}));
