+++ packages/mantine/src/components/table/table-footer/TableFooter.tsx
@@ -1,9 +1,7 @@
-import {Group, DefaultProps} from '@mantine/core';
-import {FunctionComponent, ReactNode} from 'react';
+import {Group} from '@mantine/core';
+import {FunctionComponent} from 'react';
+import {TableFooterProps} from './TableFooter.types';
 
-interface TableFooterProps extends DefaultProps {
-    children?: ReactNode;
-}
 export const TableFooter: FunctionComponent<TableFooterProps> = ({children, ...others}) => (
     <Group position="apart" px="xl" py="md" {...others}>
         {children}
