+++ packages/mantine/src/components/table/table-footer/TableFooter.types.ts
@@ -1,0 +1,6 @@
+import {DefaultProps} from '@mantine/core';
+import {ReactNode} from 'react';
+
+export interface TableFooterProps extends DefaultProps {
+    children?: ReactNode;
+}
