+++ packages/mantine/src/components/table/table-header/Th.types.ts
@@ -1,0 +1,5 @@
+import {Header} from '@tanstack/react-table';
+
+export interface ThProps<T> {
+    header: Header<T, unknown>;
+}
