+++ packages/mantine/src/components/table/table-last-updated/TableLastUpdated.styles.ts
@@ -1,0 +1,10 @@
+import {createStyles} from '@mantine/core';
+
+export default createStyles((theme) => ({
+    root: {
+        minHeight: '98px',
+    },
+    label: {
+        color: theme.colors.gray[6],
+    },
+}));
