+++ packages/mantine/src/components/table/table-last-updated/TableLastUpdated.tsx
@@ -1,28 +1,10 @@
-import {createStyles, DefaultProps, Group, Selectors, Text} from '@mantine/core';
+import {Group, Text} from '@mantine/core';
 import {useDidUpdate} from '@mantine/hooks';
 import dayjs from 'dayjs';
 import {FunctionComponent, useState} from 'react';
 import {useTable} from '../TableContext';
-
-const useStyles = createStyles((theme) => ({
-    root: {
-        minHeight: '98px',
-    },
-    label: {
-        color: theme.colors.gray[6],
-    },
-}));
-
-type TableLastUpdatedStylesNames = Selectors<typeof useStyles>;
-
-interface TableLastUpdatedProps extends DefaultProps<TableLastUpdatedStylesNames> {
-    /**
-     * Label to contextualize the date
-     *
-     * @default "Last update:"
-     */
-    label?: string;
-}
+import useStyles from './TableLastUpdated.styles';
+import {TableLastUpdatedProps} from './TableLastUpdated.types';
 
 export const TableLastUpdated: FunctionComponent<TableLastUpdatedProps & {dependencies?: never}> = ({
     label = 'Last update:',
