+++ packages/mantine/src/components/table/table-pagination/TablePagination.tsx
@@ -2,19 +2,7 @@ import {Pagination} from '@mantine/core';
 import {FunctionComponent} from 'react';
 
 import {useTable} from '../TableContext';
-
-interface TablePaginationProps {
-    /**
-     * The total number of page. Use null only if your table is paginated client side
-     */
-    totalPages: number | null;
-    /**
-     * The callback if the current page is changed.
-     *
-     * @param pageIndex The index of the updated page.
-     */
-    onPageChange?: (pageIndex: number) => void;
-}
+import {TablePaginationProps} from './TablePagination.types';
 
 export const TablePagination: FunctionComponent<TablePaginationProps> = ({totalPages, onPageChange}) => {
     const {state, setState, containerRef, getPageCount} = useTable();
