+++ packages/mantine/src/components/table/table-per-page/TablePerPage.tsx
@@ -2,27 +2,7 @@ import {Group, SegmentedControl, Text} from '@mantine/core';
 import {FunctionComponent} from 'react';
 
 import {useTable} from '../TableContext';
-
-interface TablePerPageProps {
-    /**
-     * The label displayed before the control
-     *
-     * @default Results per page
-     */
-    label?: string;
-    /**
-     * The per page choices to display
-     *
-     * @default [25, 50, 100]
-     */
-    values?: number[];
-    /**
-     * The callback if the entries per page is changed.
-     *
-     * @param perPage the new number of entries per page
-     */
-    onPerPageChange?: (perPage: number) => void;
-}
+import {TablePerPageProps} from './TablePerPage.types';
 
 export const TablePerPage: FunctionComponent<TablePerPageProps> & {DEFAULT_SIZE: number} = ({
     label = 'Results per page',
