+++ packages/mantine/src/components/table/table-predicate/TablePredicate.styles.ts
@@ -1,0 +1,8 @@
+import {createStyles} from '@mantine/core';
+
+export default createStyles((theme) => ({
+    root: {},
+    wrapper: {},
+    label: {},
+    select: {},
+}));
